Flush pending indexes periodically
During the snapshot process, we flush the index periodically in order to
commit packs frequently. If the snapshot process is interrupted, the packs
referenced by an index are not lost when resuming the snapshot. However,
flushing often means saving a lot of small indexes to file. The garbage
collector should merge all small indexes and consolidate them into bigger
index files.

+++ src/archiver/processor.rs
@@ -168,9 +168,9 @@ fn chunk_and_save_blobs(
     // same contents will no longer produce same chunks and IDs.
     let chunker = StreamCDC::with_level(
         reader,
-        global::defaults::MIN_CHUNK_SIZE,
-        global::defaults::AVG_CHUNK_SIZE,
-        global::defaults::MAX_CHUNK_SIZE,
+        global::defaults::MIN_CHUNK_SIZE as u32,
+        global::defaults::AVG_CHUNK_SIZE as u32,
+        global::defaults::MAX_CHUNK_SIZE as u32,
         Normalization::Level1,
     );
